+++ jwst/flatfield/flat_field.py
@@ -6,8 +6,6 @@ import logging
 import math
 
 import numpy as np
-import numpy.ma as ma
-from astropy.stats import sigma_clip    # for combine_fast_slow()
 
 from .. import datamodels
 from .. datamodels import dqflags
@@ -1266,6 +1264,14 @@ def read_flat_table(flat_model, exposure_type,
         tab_flat = tab_flat[filter]
     del filter1, filter2, filter
 
+    # Check that the wavelengths are increasing.  This is a requirement
+    # for using np.searchsorted (see wl_interpolate).
+    if len(tab_wl) > 1:
+        diff = tab_wl[1:] - tab_wl[0:-1]
+        if np.any(diff <= 0.):
+            log.warning("Wavelengths in the fast-variation table "
+                        "must be strictly increasing.")
+
     return (tab_wl, tab_flat)
 
 
@@ -1300,36 +1306,19 @@ def combine_fast_slow(wl, flat_2d, tab_wl, tab_flat, dispaxis):
         to the wavelengths of the science image, i.e. `wl`.
     """
 
-    (ny, nx) = wl.shape
-
-    dwl = np.zeros_like(wl)
+    wl_c = clean_wl(wl, dispaxis)
+    dwl = np.zeros_like(wl_c)
 
     if dispaxis == HORIZONTAL:
-        # The wavelength span of pixel i is (wl[i+1] - wl[i-1]) / 2.
-        temp = (wl[:, 2:] - wl[:, 0:-2]) / 2.
-        dwl[:, 1:-1] = temp
-        dwl[:, 0] = dwl[:, 1]
+        dwl[:, 0:-1] = wl_c[:, 1:] - wl_c[:, 0:-1]
         dwl[:, -1] = dwl[:, -2]
     elif dispaxis == VERTICAL:
-        temp = (wl[2:, :] - wl[0:-2, :]) / 2.
-        dwl[1:-1, :] = temp
-        dwl[0, :] = dwl[1, :]
+        dwl[0:-1, :] = wl_c[1:, :] - wl_c[0:-1, :]
         dwl[-1, :] = dwl[-2, :]
 
-    # Elements of dwl may be zero, due to values in the original
-    # wavelength array being either zero or NaN.  Create a masked array
-    # so sigma_clip will ignore those elements, and also so that the
-    # returned array will have a mask that flags outliers as well as
-    # the originally masked elements.
-    mask = (dwl == 0.)                  # NaN wavelengths were set to 0
-    dwl_ma = ma.array(dwl, mask=mask)
-    temp = sigma_clip(dwl, cenfunc="median", axis=None, masked=True)
-    replacement_value = temp.mean()
-    dwl[dwl_ma.mask] = replacement_value
-
     # Values averaged within tab_flat.
-    values = np.zeros_like(wl)
-    (ny, nx) = wl.shape
+    values = np.zeros_like(wl_c)
+    (ny, nx) = wl_c.shape
     # Abscissas and weights for 3-point Gaussian integration, but taking
     # the width of the interval to be 1, so the result will be the average
     # over the interval.
@@ -1338,11 +1327,79 @@ def combine_fast_slow(wl, flat_2d, tab_wl, tab_flat, dispaxis):
     wgt = np.array([5., 8., 5.]) / 18.
     for j in range(ny):
         for i in range(nx):
-            # average the tabular data over the range of wavelengths
-            values[j, i] = g_average(wl[j, i], dwl[j, i],
-                                     tab_wl, tab_flat, dx, wgt)
+            if wl[j, i] <= 0.:          # note:  wl, not wl_c
+                values[j, i] = 1.
+            else:
+                # average the tabular data over the range of wavelengths
+                values[j, i] = g_average(wl_c[j, i], dwl[j, i],
+                                         tab_wl, tab_flat, dx, wgt)
 
     return flat_2d * values
+
+
+def clean_wl(wl, dispaxis):
+    """Replace zeros and/or NaNs in the wl array.
+
+    Parameters
+    ----------
+    wl: 2-D ndarray
+        Wavelength at each pixel of the 2-D slit array.
+
+    dispaxis: int
+        1 is horizontal, 2 is vertical.
+
+    Returns
+    -------
+    2-D ndarray
+        A copy of `wl`, but with zero and negative values replaced with
+        an average wavelength.  For each column (row) in the dispersion
+        direction, the average to find a replacement value is taken along
+        the cross-dispersion direction.
+    """
+
+    wl_c = wl.copy()                    # so we can replace zeros
+    wl_c[wl_c <= 0.] = np.nan
+    shape = wl_c.shape
+
+    if dispaxis == HORIZONTAL:
+        i0 = None
+        i1 = None
+        for i in range(shape[1]):
+            temp = wl_c[:, i]
+            if np.any(np.isfinite(temp)):
+                if i0 is None:
+                    i0 = i              # first i with some non-zero wl
+                i1 = i                  # last i (so far) with some non-zero wl
+                replacement_value = np.nanmean(temp)
+                temp[np.isnan(temp)] = replacement_value
+        if i0 is not None and i0 > 0:
+            temp = wl_c[:, i0].reshape(shape[0], 1)
+            wl_c[:, 0:i0] = temp.copy()
+        if i1 is not None and i1 < shape[1] - 1:
+            temp = wl_c[:, i1].reshape(shape[0], 1)
+            wl_c[:, i1:] = temp.copy()
+    elif dispaxis == VERTICAL:
+        j0 = None
+        j1 = None
+        for j in range(shape[0]):
+            temp = wl_c[j, :]
+            if np.any(np.isfinite(temp)):
+                if j0 is None:
+                    j0 = j              # first j with some non-zero wl
+                j1 = j                  # last j (so far) with some non-zero wl
+                replacement_value = np.nanmean(temp)
+                temp[np.isnan(temp)] = replacement_value
+        if j0 is not None and j0 > 0:
+            temp = wl_c[j0, :].reshape(1, shape[1])
+            wl_c[0:j0, :] = temp.copy()
+        if j1 is not None and j1 < shape[0] - 1:
+            temp = wl_c[j1, :].reshape(1, shape[1])
+            wl_c[j1:, :] = temp.copy()
+    else:
+        wl_c = wl.copy()
+
+    return wl_c
+
 
 def g_average(wl0, dwl0, tab_wl, tab_flat, dx, wgt):
     """Gaussian integration.
@@ -1406,8 +1463,11 @@ def wl_interpolate(wavelength, tab_wl, tab_flat):
         The flat-field value (from `tab_flat`) at `wavelength`.
     """
 
-    if wavelength < tab_wl[0] or wavelength > tab_wl[-1]:
-        return 1.
+    if wavelength < tab_wl[0]:
+        return tab_flat[0]
+    if wavelength > tab_wl[-1]:
+        return tab_flat[-1]
+
     n0 = np.searchsorted(tab_wl, wavelength) - 1
     p = (wavelength - tab_wl[n0]) / (tab_wl[n0 + 1] - tab_wl[n0])
     q = 1. - p
@@ -1471,9 +1531,8 @@ def interpolate_flat(image_flat, image_dq, image_wl, wl):
     k = np.zeros(wl.shape, dtype=np.intp) - 1
 
     # Truncate the index for wavelengths that are outside the range of
-    # image_wl.  Near the end of this function we'll flag these pixels
-    # with NO_FLAT_FIELD and set flat_2d to 1, but for now the indices
-    # need to be assigned harmless values to avoid indexing out of bounds.
+    # image_wl.  The indices need to be assigned harmless values to avoid
+    # indexing out of bounds.
     #   Why do we set the upper limit of k to nz - 2?
     #   Because we interpolate using elements k and k + 1.
     k[:, :] = np.where(wl <= image_wl[0], 0, k)
@@ -1507,11 +1566,9 @@ def interpolate_flat(image_flat, image_dq, image_wl, wl):
                            np.bitwise_or(image_dq[k, iypixel, ixpixel],
                                          image_dq[k + 1, iypixel, ixpixel]))
 
-    # If the wavelength is out of range, set a DQ flag.
-    indx = np.where(wl < image_wl[0])
-    flat_dq[indx] |= dqflags.pixel['NO_FLAT_FIELD']
-    indx = np.where(wl > image_wl[nz - 1])
-    flat_dq[indx] |= dqflags.pixel['NO_FLAT_FIELD']
+    # If the wavelength at a pixel is zero, the flat field cannot be
+    # determined.
+    flat_dq[wl <= 0.] |= dqflags.pixel['NO_FLAT_FIELD']
 
     # If a pixel is flagged as bad, applying flat_2d should not make any
     # change to the science data.